Revert "Restaurant view 1.0"
