+++ TravelManagePlanner/TravelManagePlanner/HomeTabSB.storyboard
@@ -51,75 +51,7 @@
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="IjG-38-8qV" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="1810" y="699"/>
-        </scene>
-        <!--Recommend Page View Controller-->
-        <scene sceneID="Eiq-dt-gR6">
-            <objects>
-                <pageViewController storyboardIdentifier="RecommendPageViewSB" autoresizesArchivedViewToFullSize="NO" transitionStyle="scroll" navigationOrientation="horizontal" spineLocation="none" id="wcS-jK-2Bb" customClass="RecommendPageViewController" customModule="TravelManagePlanner" customModuleProvider="target" sceneMemberID="viewController"/>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="IHV-Z2-DZy" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="2532" y="1"/>
-        </scene>
-        <!--Restaurant View Controller-->
-        <scene sceneID="Fce-Np-bbY">
-            <objects>
-                <viewController storyboardIdentifier="RestaurantViewSB" id="aic-py-dHX" customClass="RestaurantViewController" customModule="TravelManagePlanner" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="svt-55-ESL">
-                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <viewLayoutGuide key="safeArea" id="hNl-LC-GMf"/>
-                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-                    </view>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="pyT-eo-bIo" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="2532" y="699"/>
-        </scene>
-        <!--Tour Area View Controller-->
-        <scene sceneID="WlL-bU-QbP">
-            <objects>
-                <viewController storyboardIdentifier="TourAreaViewSB" useStoryboardIdentifierAsRestorationIdentifier="YES" id="aQK-qr-mtY" customClass="TourAreaViewController" customModule="TravelManagePlanner" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="eMD-5U-KoN">
-                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <viewLayoutGuide key="safeArea" id="GIr-sp-GgV"/>
-                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-                    </view>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="F2N-P6-kIO" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="3267" y="698"/>
-        </scene>
-        <!--Restaurant Calendar View Controller-->
-        <scene sceneID="sty-zT-VCh">
-            <objects>
-                <viewController storyboardIdentifier="RestaurantCalendarViewSB" id="aKT-vK-eoS" customClass="RestaurantCalendarViewController" customModule="TravelManagePlanner" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="g47-y3-Aiv">
-                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <viewLayoutGuide key="safeArea" id="cfx-2X-QeO"/>
-                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-                    </view>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="ExW-0u-due" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="2532" y="1348"/>
-        </scene>
-        <!--View Controller-->
-        <scene sceneID="qtn-DR-Bx0">
-            <objects>
-                <viewController id="Ipp-iS-xmr" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="x4F-Ht-bLs">
-                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <viewLayoutGuide key="safeArea" id="8xM-cs-1rC"/>
-                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-                    </view>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="0Zd-MX-V0N" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="3267" y="1348"/>
+            <point key="canvasLocation" x="2113" y="1"/>
         </scene>
         <!--Desti Search View Controller-->
         <scene sceneID="n2x-Se-xLn">
@@ -136,10 +68,10 @@
             </objects>
             <point key="canvasLocation" x="1357" y="1"/>
         </scene>
-        <!--Accomo Calendar View Controller-->
+        <!--Calendar Modal View Controller-->
         <scene sceneID="GoP-7B-LhI">
             <objects>
-                <viewController storyboardIdentifier="AccomoCalendarViewSB" id="Yde-2N-gR5" customClass="AccomoCalendarViewController" customModule="TravelManagePlanner" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="CalendarModalViewSB" id="Yde-2N-gR5" customClass="CalendarModalViewController" customModule="TravelManagePlanner" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="bRN-Ux-ivS">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -149,7 +81,7 @@
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="zGR-nN-Huf" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="1810" y="1347"/>
+            <point key="canvasLocation" x="2112" y="650"/>
         </scene>
     </scenes>
     <resources>
