--- TravelManagePlanner/TravelManagePlanner/Model/HomeTabModel/GlobalHomeTab.swift
@@ -1,16 +1,0 @@
-//
-//  HomeTabStruct.swift
-//  TravelManagePlanner
-//
-//  Created by YoonDaeSung on 2022/02/09.
-//
-
-//import Foundation
-//
-//
-//struct GlobalHomeTab {
-//    
-//    // 홈탭 데이터
-//    static var globalHomeTabData = HomeTabData(themeData: "", dayToGocalendar: "", dayToComecalendar: "", numPeople:"", budgetAmount: "")
-//    
-//}
