+++ TravelManagePlanner/TravelManagePlanner/View/HomeTabView/AccomoViewCell.swift
@@ -6,7 +6,13 @@
 //
 import UIKit
 
+// 범용성을 위해 class가 아닌 AnyObject로 선언해준다.
+protocol ContentsMainTextDelegate: AnyObject {
 
+// 위임해줄 기능
+func categoryButtonTapped()
+    
+}
 
 
 class AccomoViewCell: UITableViewCell {
@@ -94,6 +100,9 @@ class AccomoViewCell: UITableViewCell {
     }
     
     func setUpView() {
+//        addSubview(accomoImgButton)
+//        addSubview(accomoAllTitleStack)
+        
         contentView.addSubview(accomoImgButton)
         contentView.addSubview(accomoAllTitleStack)
