+++ TravelManagePlanner/TravelManagePlanner/ViewController/HomeTabVC/DestiSearchViewController.swift
@@ -10,9 +10,6 @@ import Alamofire
 // 검색기능
 class DestiSearchViewController: UIViewController {
     
-    //뷰모델 소유
-    var homeTabViewModel = HomeTabViewModel()
-    
     // 테마 타이틀
     lazy var themeTitleLabel: UILabel = {
         let label = UILabel()
@@ -114,15 +111,12 @@ extension DestiSearchViewController: UITableViewDelegate {
         
         print("테이블뷰 클릭")
         // main에 있는 두번째화면 불러오기(스토리보드 활용)
-        let nextView = UIStoryboard(name: "HomeTabSB", bundle: nil).instantiateViewController(withIdentifier: "RecommendPageViewSB") as! RecommendPageViewController
+        let nextView = UIStoryboard(name: "HomeTabSB", bundle: nil).instantiateViewController(withIdentifier: "AccomoViewSB") as! AccomoViewController
         
         // 다음화면에서 바텀탭 없애기
+//        nextView.hidesBottomBarWhenPushed = true
         navigationController!.pushViewController(nextView, animated: true)
-   
-        let placeData = self.data.placeData[indexPath.row]
-        homeTabViewModel.updateDestiSearchData(userDestiData: placeData)
         
-        print("선택된 장소 :", placeData)
     }
 }
