+++ TravelManagePlanner/TravelManagePlanner/ViewController/HomeTabVC/ThemeViewController.swift
@@ -10,17 +10,14 @@ import UIKit
 // 커밋 테스트
 class ThemeViewController: UIViewController{
     
-    // 뷰모델 소지
-    var homeTabViewModel = HomeTabViewModel()
-    
     let cellID = "Cell"
     // Assets의 사진 출력
-    lazy var imgDataList = ["커플", "가족", "우정", "기타"]
+    lazy var imgDataName = ["커플", "가족", "우정", "기타"]
     
     var imgArray: [UIImage] {
         var img:[UIImage] = []
         
-        for i in imgDataList {
+        for i in imgDataName {
             if let asImg = UIImage(named: i) {
                 
                 img.append(asImg)
@@ -69,6 +66,8 @@ class ThemeViewController: UIViewController{
         setLayout()
         setDelegate()
     }
+    
+    
     
     // 뷰 요소들 화면에 세팅하기
     func setUpView() {
@@ -125,12 +124,13 @@ class ThemeViewController: UIViewController{
     }
     
 } // class
+
 extension ThemeViewController: UICollectionViewDataSource {
     
     // 셀 갯수 설정(필수)
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         
-        return imgDataList.count
+        return imgDataName.count
     }
     
     // 셀 데이터 내용 불러오기
@@ -139,10 +139,20 @@ extension ThemeViewController: UICollectionViewDataSource {
         let cell = collectionView.dequeueReusableCell(withReuseIdentifier: cellID, for: indexPath) as! ThemeViewCell
         
         //      cell.imageView.image = (imgArray[indexPath.row])
-        cell.cellLoadImage(imgDataList[indexPath.row])
-        cell.imgLabel.text = imgDataList[indexPath.row]
+        cell.cellLoadImage(imgDataName[indexPath.row])
+        cell.imgLabel.text = imgDataName[indexPath.row]
         
         return cell
+    }
+    
+    // 해당 셀 선택시에 액션
+    func collectionView(collectionView: UICollectionView, didSelectItemAtIndexPath indexPath: NSIndexPath) {
+        
+        //해당 셀클릭시 디테일 버튼 추후 구현하기
+        //collectionView.cellForItemAtIndexPath(1)?.backgroundColor = UIColor.grayColor
+        
+        // 내가고른게 선택되는지 찍어보기
+        print("클릭됬니")
     }
 }
 
@@ -152,22 +162,10 @@ extension ThemeViewController: UICollectionViewDelegate {
         
         // main에 있는 두번째화면 불러오기(스토리보드 활용)
         let nextView = UIStoryboard(name: "HomeTabSB", bundle: nil).instantiateViewController(withIdentifier: "DetaileSettingViewSB") as! DetaileSettingViewController
-        homeTabViewModel.register()
-        print("register 버튼 클릭")
-        print(indexPath)
         
         // 다음화면에서 바텀탭 없애기
         nextView.hidesBottomBarWhenPushed = true
         navigationController!.pushViewController(nextView, animated: true)
-        
-        let imgDataName = imgDataList[indexPath.row]
-        
-        // 선택한 테마 데이터 넘기기
-        homeTabViewModel.updateThemeData(userThemeData: imgDataName)
-//        nextView.homeTabViewModel.userThemeData = imgDataName
-        // 선택한 데이터 넘기기
-       
-//        homeTabViewModel.themeData = imgDataName
     }
 }
